weko#44312 Fix activity lock

+++ modules/weko-workflow/weko_workflow/utils.py
@@ -4485,7 +4485,7 @@ def delete_user_lock_activity_cache(activity_id, data):
     cur_locked_val = str(get_cache_data(cache_key)) or str()
     msg = _("Not unlock")
 
-    if cur_locked_val and not data["is_opened"] or (cur_locked_val == activity_id):
+    if cur_locked_val and not data["is_opened"] or (cur_locked_val == activity_id and data['is_force']):
         delete_cache_data(cache_key)
         msg = "User Unlock Success"
     return msg
